Add some custom join query methods
This app was suffering from the well-known N+1 select problem where a
first query is done once but from each row returned is performed another
request to complete the data required.

This conflicts with ORM (Object-relational mapping) that is the desired
way of working in objects. The suggested solution finding ways to use
JOINs or VIEW when quering. Following that suggestion this patch implements
3 different methods to boost up the query speed.